+++ src/renderer/routes/Settings/components/CustomRPCModal/CustomRPCItem.tsx
@@ -1,9 +1,8 @@
+import { useMemo, useState } from 'react';
 import { RPCItem } from '@/isomorphic/types/rabbyx';
 import ChainIcon from '@/renderer/components/ChainIcon';
-import { DeleteWrapper } from '@/renderer/components/DeleteWrapper';
 import { Switch } from '@/renderer/components/Switch/Switch';
 import { findChain } from '@/renderer/utils/chain';
-import { useMemo, useState } from 'react';
 import styles from './index.module.less';
 
 interface CustomtRPCItemProps {
@@ -26,49 +25,36 @@ export const CustomtRPCItem = (props: CustomtRPCItemProps) => {
   }, [data.id]);
 
   return (
-    <DeleteWrapper
-      timeout={0}
-      className={styles.part}
-      onCancelDelete={() => {
-        setIsShowCheck(false);
-      }}
-      onConfirmDelete={() => {
-        onDelete?.(data?.id);
-        setIsShowCheck(false);
-      }}
-      showClose={isShowCheck}
-    >
-      <div className={styles.customRPCItem}>
-        <div className={styles.switch}>
-          <Switch
-            checked={data?.rpc?.enable}
-            onChange={(v) => onEnable?.(data.id, v)}
-          />
-        </div>
-        <ChainIcon chain={data?.id} isShowCustomRPC className={styles.icon} />
-        <div className={styles.rowContent}>
-          <div className={styles.name}>{chain?.name}</div>
-          <div className={styles.url} title={data?.rpc?.url}>
-            {data?.rpc?.url}
-          </div>
-        </div>
-        <div className={styles.action}>
-          <img
-            src="rabby-internal://assets/icons/mainwin-settings/icon-edit.svg"
-            alt=""
-            onClick={() => {
-              onEdit?.(data);
-            }}
-          />
-          <img
-            src="rabby-internal://assets/icons/mainwin-settings/icon-delete.svg"
-            alt=""
-            onClick={() => {
-              setIsShowCheck(true);
-            }}
-          />
+    <div className={styles.customRPCItem}>
+      <div className={styles.switch}>
+        <Switch
+          checked={data?.rpc?.enable}
+          onChange={(v) => onEnable?.(data.id, v)}
+        />
+      </div>
+      <ChainIcon chain={data?.id} isShowCustomRPC className={styles.icon} />
+      <div className={styles.rowContent}>
+        <div className={styles.name}>{chain?.name}</div>
+        <div className={styles.url} title={data?.rpc?.url}>
+          {data?.rpc?.url}
         </div>
       </div>
-    </DeleteWrapper>
+      <div className={styles.action}>
+        <img
+          src="rabby-internal://assets/icons/mainwin-settings/icon-edit.svg"
+          alt=""
+          onClick={() => {
+            onEdit?.(data);
+          }}
+        />
+        <img
+          src="rabby-internal://assets/icons/mainwin-settings/icon-delete.svg"
+          alt=""
+          onClick={() => {
+            onDelete?.(data?.id);
+          }}
+        />
+      </div>
+    </div>
   );
 };