+++ src/renderer/utils/chain.ts
@@ -1,6 +1,9 @@
 import defaultSuppordChain from '@/isomorphic/default-support-chains.json';
 import { Chain } from '@debank/common';
-import { SupportedChain } from '@rabby-wallet/rabby-api/dist/types';
+import { SupportedChain, TokenItem } from '@rabby-wallet/rabby-api/dist/types';
+import { TestnetChain } from '@/isomorphic/types/customTestnet';
+import { CustomTestnetToken } from '@/isomorphic/types/rabbyx';
+import BigNumber from 'bignumber.js';
 import { intToHex } from './number';
 
 export function supportedChainToChain(item: SupportedChain): Chain {
@@ -123,42 +126,8 @@ const store = {
     .map((item) => {
       return supportedChainToChain(item);
     }),
-  testnetList: [],
+  testnetList: [] as TestnetChain[],
 };
-
-export const updateChainStore = (params: Partial<typeof store>) => {
-  Object.assign(store, params);
-};
-
-export const findChain = (params: {
-  enum?: CHAINS_ENUM | string | null;
-  id?: number | null;
-  serverId?: string | null;
-  hex?: string | null;
-  networkId?: string | null;
-  name?: string | null;
-}) => {
-  const { enum: chainEnum, id, serverId, hex, networkId, name } = params;
-  // if (chainEnum && chainEnum.startsWith('CUSTOM_')) {
-  //   return findChain({
-  //     id: +chainEnum.replace('CUSTOM_', ''),
-  //   });
-  // }
-  const chain = [...store.mainnetList].find(
-    (item) =>
-      item.enum === chainEnum ||
-      (id && +item.id === +id) ||
-      item.serverId === serverId ||
-      item.hex === hex ||
-      item.network === networkId ||
-      item.name === name
-  );
-  return chain;
-};
-
-export const DEFAULT_ETH_CHAIN = findChain({
-  enum: 'ETH',
-})!;
 
 export const getChainList = (net?: 'mainnet' | 'testnet') => {
   if (net === 'mainnet') {
@@ -169,6 +138,50 @@ export const getChainList = (net?: 'mainnet' | 'testnet') => {
   }
   return [...store.mainnetList, ...store.testnetList];
 };
+
+export const updateChainStore = (params: Partial<typeof store>) => {
+  Object.assign(store, params);
+};
+
+export const findChain = (
+  params: {
+    enum?: CHAINS_ENUM | string | null;
+    id?: number | null;
+    serverId?: string | null;
+    hex?: string | null;
+    networkId?: string | null;
+    name?: string | null;
+  },
+  _chainList?: (Chain | TestnetChain)[]
+): Chain | TestnetChain | null | undefined => {
+  const chainList = _chainList || [
+    ...getChainList('mainnet'),
+    ...getChainList('testnet'),
+  ];
+  const { enum: chainEnum, id, serverId, hex, networkId, name } = params;
+  if (chainEnum && chainEnum.startsWith('CUSTOM_')) {
+    return findChain(
+      {
+        id: +chainEnum.replace('CUSTOM_', ''),
+      },
+      _chainList
+    );
+  }
+  const chain = chainList.find(
+    (item) =>
+      item.enum === chainEnum ||
+      (id && +item.id === +id) ||
+      item.serverId === serverId ||
+      item.hex === hex ||
+      item.network === networkId
+  );
+
+  return chain;
+};
+
+export const DEFAULT_ETH_CHAIN = findChain({
+  enum: 'ETH',
+})!;
 
 /**
  * @description safe find chain
@@ -190,3 +203,32 @@ export function findChainByServerID(chainId: Chain['serverId']): Chain | null {
 
 // export { formatChain } from '@/isomorphic/wallet/chain';
 // export type { DisplayChainWithWhiteLogo } from '@/isomorphic/wallet/chain';
+
+export const customTestnetTokenToTokenItem = (
+  token: CustomTestnetToken
+): TokenItem => {
+  const chain = findChain({
+    id: token.chainId,
+  });
+  return {
+    id: token.id,
+    chain: chain?.serverId || '',
+    amount: token.amount,
+    raw_amount: token.rawAmount,
+    raw_amount_hex_str: `0x${new BigNumber(token.rawAmount || 0).toString(16)}`,
+    decimals: token.decimals,
+    display_symbol: token.symbol,
+    is_core: false,
+    is_verified: false,
+    is_wallet: false,
+    is_scam: false,
+    is_suspicious: false,
+    logo_url: '',
+    name: token.symbol,
+    optimized_symbol: token.symbol,
+    price: 0,
+    symbol: token.symbol,
+    time_at: 0,
+    price_24h_change: 0,
+  };
+};